ref(dropdownControl): Remove V1
