+++ static/app/views/dashboardsV2/widgetCard/widgetCardContextMenu.tsx
@@ -6,8 +6,8 @@ import {openDashboardWidgetQuerySelectorModal} from 'sentry/actionCreators/modal
 import Feature from 'sentry/components/acl/feature';
 import Button from 'sentry/components/button';
 import {openConfirmModal} from 'sentry/components/confirm';
-import DropdownMenuControlV2 from 'sentry/components/dropdownMenuControlV2';
-import {MenuItemProps} from 'sentry/components/dropdownMenuItemV2';
+import DropdownMenuControlV2 from 'sentry/components/dropdownMenuControl';
+import {MenuItemProps} from 'sentry/components/dropdownMenuItem';
 import {isWidgetViewerPath} from 'sentry/components/modals/widgetViewerModal/utils';
 import Tag from 'sentry/components/tag';
 import Tooltip from 'sentry/components/tooltip';
