+++ static/app/views/performance/landing/widgets/components/widgetContainer.tsx
@@ -4,7 +4,7 @@ import omit from 'lodash/omit';
 import pick from 'lodash/pick';
 import * as qs from 'query-string';
 
-import DropdownButtonV2 from 'sentry/components/dropdownButtonV2';
+import DropdownButton from 'sentry/components/dropdownButton';
 import CompositeSelect from 'sentry/components/forms/compositeSelect';
 import {IconEllipsis} from 'sentry/icons/iconEllipsis';
 import {t} from 'sentry/locale';
@@ -186,7 +186,7 @@ export const WidgetContainerActions = ({
 
   function trigger({props, ref}) {
     return (
-      <DropdownButtonV2
+      <DropdownButton
         ref={ref}
         {...props}
         size="xs"
